Settings variable now accepts 32bit values

+++ grbl/system.c
@@ -123,7 +123,7 @@ void system_execute_startup(char *line)
 uint8_t system_execute_line(char *line)
 {
   uint8_t char_counter = 1;
-  uint8_t helper_var = 0; // Helper variable
+  uint32_t helper_var = 0; // Helper variable
   float parameter, value;
   switch( line[char_counter] ) {
     case 0 : report_grbl_help(); break;
